[refactor] Extract helpers to extension files
